Renaming bad called module, fix imports
