+++ tests/generic_views/views.py
@@ -6,8 +6,8 @@ from django.core.urlresolvers import reverse, reverse_lazy
 from django.utils.decorators import method_decorator
 from django.views import generic
 
+from .forms import AuthorForm, ContactForm
 from .models import Artist, Author, Book, BookSigning, Page
-from .test_forms import AuthorForm, ContactForm
 
 
 class CustomTemplateView(generic.TemplateView):
